+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## Version 3.2.1
 
-- prevent Text-to-Speech announcements for boss units during the Lothraxion encounter
+- Lothraxion will no longer spam Text-To-Speech announcements for Lingering Image casts
 - fix a bug where some spells with inconsistent events such as Leech Veil would persist as bar elements past the end of their cast time